refactor(cli): src/cli.ts → cli/index.ts + 抽 parse-run-config / progress / update-check / coverage-renderer
修 sibling-folder 反模式(src/cli.ts 与 src/cli/ 同级)、cli.ts 1956 行单文件过大、analysis/sample-design-coverage-cli.ts 错位(CLI 渲染放在 analysis dir)三个问题。

变化:
- src/cli.ts → src/cli/index.ts(package.json bin / postbuild / test/cli.test.ts CLI path 同步)
- 抽 parseRunConfig + RUN_OPTIONS / COMMON_OPTIONS / DEFAULT_REPORTS_DIR → src/cli/parse-run-config.ts(269 行)
- 抽 makeOnProgress + ProgressInfo type → src/cli/progress.ts(90 行)
- 抽 checkUpdate → src/cli/update-check.ts(21 行)
- src/analysis/sample-design-coverage-cli.ts → src/cli/coverage-renderer.ts(本质 CLI 渲染,跟 src/renderer/ HTML 渲染语义对齐;函数名 renderSampleDesignCoverage 不动)
- test/analysis/sample-design-coverage-cli.test.ts → test/cli/coverage-renderer.test.ts

效果:
- src/cli/index.ts:1956 → 1583 行(-373)
- 纯 file move + 路径调整,不动逻辑;809 测试全过,0 lint errors,bin smoke-test ok

阶段 1 结束;阶段 2(单 handler 超 250 行时拆 commands/<name>.ts)按改动驱动,本 PR 不做。
